+++ CHANGELOG.md
@@ -1,3 +1,31 @@
+# [2.21.0](https://github.com/UCLALibrary/ucla-library-website-components/compare/v2.20.0...v2.21.0) (2022-12-20)
+
+
+### Features
+
+* search components QA ([#283](https://github.com/UCLALibrary/ucla-library-website-components/issues/283)) ([db35999](https://github.com/UCLALibrary/ucla-library-website-components/commit/db35999de75558bc4111f11019e99de6cbde428a))
+
+# [2.20.0](https://github.com/UCLALibrary/ucla-library-website-components/compare/v2.19.1...v2.20.0) (2022-12-20)
+
+
+### Features
+
+* APPS-2106 Add Remove filters to search-generic ([#282](https://github.com/UCLALibrary/ucla-library-website-components/issues/282)) ([3588f20](https://github.com/UCLALibrary/ucla-library-website-components/commit/3588f20534962e83ffacf5243e4d74940e84dec9))
+
+## [2.19.1](https://github.com/UCLALibrary/ucla-library-website-components/compare/v2.19.0...v2.19.1) (2022-12-16)
+
+
+### Bug Fixes
+
+* APPS-2108 Update sectionSummary in BlockMediaGalleryExtraFieldsFragment query ([#281](https://github.com/UCLALibrary/ucla-library-website-components/issues/281)) ([947339a](https://github.com/UCLALibrary/ucla-library-website-components/commit/947339a3f0076d1ece176eccbdf29c32c5bb46a5))
+
+# [2.19.0](https://github.com/UCLALibrary/ucla-library-website-components/compare/v2.18.0...v2.19.0) (2022-12-14)
+
+
+### Features
+
+* location detail component QA ([#278](https://github.com/UCLALibrary/ucla-library-website-components/issues/278)) ([62a3db6](https://github.com/UCLALibrary/ucla-library-website-components/commit/62a3db61a771ad6f3cadcf10c8f4067bc880f09b))
+
 # [2.18.0](https://github.com/UCLALibrary/ucla-library-website-components/compare/v2.17.1...v2.18.0) (2022-12-14)